Move newshape and expand hypothesis tests into their own files
This removes duplication for the newshape tests. The integer and slice
exhaustive tests are still in the respective files for the types.
