+++ ndindex/tests/test_booleanarray.py
@@ -88,46 +88,6 @@ def test_booleanarray_reduce_hypothesis(idx, shape):
         # give an IndexError
         assert reduced == index
 
-@example(array([[[True], [False]]]), (1, 1, 2))
-@example(full((1, 9), False), (3, 3))
-@given(boolean_arrays, one_of(shapes, integers(0, 10)))
-def test_booleanarray_newshape_hypothesis(idx, shape):
-    if isinstance(shape, int):
-        a = arange(shape)
-    else:
-        a = arange(prod(shape)).reshape(shape)
-
-    def raw_func(a, idx):
-        return a[idx].shape
-
-    def ndindex_func(a, index):
-        return index.newshape(shape)
-
-    def assert_equal(raw_shape, newshape):
-        assert raw_shape == newshape
-
-    index = BooleanArray(idx)
-    if (index.count_nonzero == 0
-        and a.shape != index.shape
-        and prod(a.shape) == prod(index.shape)
-        and any(i != 0 and i != j for i, j in zip(index.shape, a.shape))
-        and len(a.shape) == len(index.shape)):
-        # NumPy currently allows this case, due to a bug: (see
-        # https://github.com/numpy/numpy/issues/16997 and
-        # https://github.com/numpy/numpy/pull/17010), but we disallow it.
-        with raises(IndexError, match=r"boolean index did not match indexed "
-                    r"array along dimension \d+; dimension is \d+ but "
-                    r"corresponding boolean dimension is \d+"):
-            index.newshape(shape)
-        # Make sure this really is one of the cases NumPy lets through. Remove
-        # this once a version of NumPy is released with the above fix.
-        a[index.raw]
-        return
-
-    check_same(a, idx, raw_func=raw_func, ndindex_func=ndindex_func,
-               assert_equal=assert_equal, same_exception=False)
-
-
 @given(boolean_arrays, one_of(shapes, integers(0, 10)))
 def test_booleanarray_isempty_hypothesis(idx, shape):
     if isinstance(shape, int):
