+++ ndindex/tests/test_newaxis.py
@@ -1,14 +1,9 @@
 from numpy import arange, newaxis
 
-from pytest import raises
-
 from hypothesis import given
 from hypothesis.strategies import one_of, integers
 
 from ..ndindex import ndindex
-from ..integer import Integer
-from ..tuple import Tuple
-from ..newaxis import Newaxis
 from .helpers import check_same, prod, shapes, newaxes
 
 def test_newaxis_exhaustive():
@@ -44,29 +39,6 @@ def test_newaxis_reduce_no_shape_exhaustive():
 def test_newaxis_reduce_no_shape_hypothesis(idx, shape):
     a = arange(prod(shape)).reshape(shape)
     check_same(a, idx, ndindex_func=lambda a, x: a[x.reduce().raw])
-
-@given(newaxes(), one_of(shapes, integers(0, 10)))
-def test_newaxis_newshape_hypothesis(idx, shape):
-    if isinstance(shape, int):
-        a = arange(shape)
-    else:
-        a = arange(prod(shape)).reshape(shape)
-
-    def raw_func(a, idx):
-        return a[idx].shape
-
-    def ndindex_func(a, index):
-        return index.newshape(shape)
-
-    def assert_equal(raw_shape, newshape):
-        assert raw_shape == newshape
-
-    check_same(a, idx, raw_func=raw_func, ndindex_func=ndindex_func,
-               assert_equal=assert_equal, same_exception=False)
-
-def test_newaxis_newshape_ndindex_input():
-    raises(TypeError, lambda: Newaxis().newshape(Tuple(2, 1)))
-    raises(TypeError, lambda: Newaxis().newshape(Integer(2)))
 
 @given(newaxes(), one_of(shapes, integers(0, 10)))
 def test_newaxis_isempty_hypothesis(idx, shape):
